+++ src/components/settings/deactivate.jsx
@@ -1,9 +1,13 @@
 import { useState, useCallback, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import cn from 'classnames';
+import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 
 import { Throbber } from '../throbber';
 import { doSequence } from '../../redux/async-helpers';
+import { pluralForm } from '../../utils';
+import UserName from '../user-name';
+import { Icon } from '../fontawesome-icons';
 import { SettingsPage } from './layout';
 import { suspendMe, unauthenticated } from './../../redux/action-creators';
 import styles from './settings.module.scss';
@@ -59,6 +63,7 @@ export default function PrivacyPage() {
         </p>
         <p>We&#x2019;re sorry to see you go.</p>
       </section>
+      <OrphanGroupsWarning />
       <section className={styles.formSection}>
         <form onSubmit={submit}>
           <div className="form-group">
@@ -88,5 +93,55 @@ export default function PrivacyPage() {
         </form>
       </section>
     </SettingsPage>
+  );
+}
+
+function OrphanGroupsWarning() {
+  const managedGroups = useSelector((state) => state.managedGroups);
+  const orphanGroups = managedGroups.filter(
+    (g) => g.administrators.length === 1 && g.isRestricted === '0',
+  );
+
+  if (orphanGroups.length === 0) {
+    return null;
+  }
+
+  if (orphanGroups.length === 1) {
+    const [g] = orphanGroups;
+    return (
+      <section className={styles.formSection}>
+        <p>
+          <Icon icon={faExclamationTriangle} /> Please note: there is a non-restricted group{' '}
+          <UserName user={g}>{g.screenName}</UserName> in which you are the only administrator.
+        </p>
+        <p>
+          If you delete your account, this group will become restricted and no one will be able to
+          create posts in it. If you don&#x2019;t want that, assign additional administrators to
+          this group before you delete the account.
+        </p>
+      </section>
+    );
+  }
+
+  return (
+    <section className={styles.formSection}>
+      <p>
+        <Icon icon={faExclamationTriangle} /> Please note: there are{' '}
+        {pluralForm(orphanGroups.length, 'non-restricted group')} in which you are the only
+        administrator:
+      </p>
+      <ol>
+        {orphanGroups.map((g) => (
+          <li key={g.id}>
+            <UserName user={g}>{g.screenName}</UserName>
+          </li>
+        ))}
+      </ol>
+      <p>
+        If you delete your account, these groups will become restricted and no one will be able to
+        create posts in them. If you don&#x2019;t want that, assign additional administrators to
+        those groups before you delete the account.
+      </p>
+    </section>
   );
 }