+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.100.0] - Not released
+### Added
+- Warning about the orphan groups on the "Delete account" page.
+
 ### Changed
 - The Highcharts library replaced by the Recharts due to license incompatibility.